Add files via upload
GUI for Sudoku has been created and an example grid has been provided.. The game is playable, but functionality is currently limited to the bare basics: inputting and deleting numbers. It is impossible for the player to delete the numbers initially provided, as deleting something and forgetting what was there might make the puzzle non-unique, which would then not be a valid Sudoku.